pyproject.toml
[project]
name = "psaiops"
version = "0.6.7"
description = "Web apps to inspect & engineer NN activations."
authors = [{ name = "apehex", email = "17317183+apehex@users.noreply.github.com" }]
readme = ".github/README.md"
license = "AGPL-3.0"
license-files = [".github/LICENSE.md"]
requires-python = ">=3.10, <3.14"
dependencies = [
    "accelerate>=1.10",
    "deformers>=0.0",
    "gradio>=5.0",
    "kernels>=0.10",
    "matplotlib>=3.10",
    "mlable-torch>=0.2",
    "requests>=2.0",
    "triton==3.4",
]

[dependency-groups]
dev = ["pytest"]

[tool.hatch.build.targets.sdist]
include = ["psaiops"]

[tool.hatch.build.targets.wheel]
include = ["psaiops"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
